Migrate /trees/new to be just /trees, keep what we need from leftover
