+++ apps/docs/app/trees/DemoFlatten.tsx
@@ -9,12 +9,12 @@ import {
 import Link from 'next/link';
 import type { CSSProperties } from 'react';
 
-import { TreeExampleHeading } from '../../components/TreeExampleHeading';
-import { FeatureHeader } from '../../diff-examples/FeatureHeader';
-import { sampleFileList } from '../demo-data';
-import { DEFAULT_FILE_TREE_PANEL_CLASS } from '../tree-examples/demo-data';
-import { TreeExampleSection } from '../tree-examples/TreeExampleSection';
+import { TreeExampleHeading } from '../components/TreeExampleHeading';
+import { FeatureHeader } from '../diff-examples/FeatureHeader';
+import { sampleFileList } from './demo-data';
 import { TREE_NEW_VIEWPORT_HEIGHTS } from './dimensions';
+import { DEFAULT_FILE_TREE_PANEL_CLASS } from './tree-examples/demo-data';
+import { TreeExampleSection } from './tree-examples/TreeExampleSection';
 import { PRODUCTS } from '@/app/product-config';
 import type { FileTreePathOptions } from '@/lib/fileTreePathOptions';
