+++ apps/docs/app/trees/DemoMiddotTruncationClient.tsx
@@ -13,10 +13,10 @@ import type {
 } from 'react';
 import { useCallback, useEffect, useRef, useState } from 'react';
 
-import { FeatureHeader } from '../../diff-examples/FeatureHeader';
-import { DEFAULT_FILE_TREE_PANEL_CLASS } from '../tree-examples/demo-data';
-import { TreeExampleSection } from '../tree-examples/TreeExampleSection';
+import { FeatureHeader } from '../diff-examples/FeatureHeader';
 import { TREE_NEW_VIEWPORT_HEIGHTS } from './dimensions';
+import { DEFAULT_FILE_TREE_PANEL_CLASS } from './tree-examples/demo-data';
+import { TreeExampleSection } from './tree-examples/TreeExampleSection';
 import { PRODUCTS } from '@/app/product-config';
 import { ButtonGroup, ButtonGroupItem } from '@/components/ui/button-group';
