+++ apps/docs/app/trees/tree-examples/demo-data.ts
@@ -1,11 +1,6 @@
 import type { CSSProperties } from 'react';
 
-import { sharedDemoFileTreeOptions } from '../../trees/demo-data';
-import type {
-  FileTreeOptions,
-  FileTreeSearchMode,
-  GitStatusEntry,
-} from '@/lib/treesCompat';
+import type { GitStatusEntry } from '@/lib/treesCompat';
 
 /** Default panel look for FileTree in docs examples. Apply via className + style on FileTree. */
 export const DEFAULT_FILE_TREE_PANEL_CLASS =
@@ -20,38 +15,3 @@ export const GIT_STATUSES_A: GitStatusEntry[] = [
   { path: 'src/components/Button.tsx', status: 'added' },
   { path: '.gitignore', status: 'deleted' },
 ];
-
-export const GIT_STATUSES_B: GitStatusEntry[] = [
-  { path: 'README.md', status: 'modified' },
-  { path: 'src/lib/utils.ts', status: 'modified' },
-  { path: 'src/utils/worker.ts', status: 'added' },
-];
-
-/** Options with flatten empty directories enabled (nested folders collapsed). Pass initialExpandedItems on the component for initial open folders (e.g. ['build']). */
-export function flatteningOptions(flatten: boolean): FileTreeOptions {
-  return {
-    ...sharedDemoFileTreeOptions,
-    flattenEmptyDirectories: flatten,
-  };
-}
-
-/** Base options for all tree example sections. */
-export const baseTreeOptions = sharedDemoFileTreeOptions;
-
-/** Options for drag-and-drop examples. Optional lockedPaths prevents those paths from being dragged. */
-export function dragDropOptions(lockedPaths?: string[]): FileTreeOptions {
-  return {
-    ...baseTreeOptions,
-    dragAndDrop: true,
-    ...(lockedPaths != null && lockedPaths.length > 0 && { lockedPaths }),
-  };
-}
-
-/** Options with search mode for the search example. Pass fileTreeSearchMode at top level so the tree applies it. Use stateConfig.initialSearchQuery in the component for prepopulated search. */
-export function searchOptions(mode: FileTreeSearchMode): FileTreeOptions {
-  return {
-    ...sharedDemoFileTreeOptions,
-    fileTreeSearchMode: mode,
-    search: true,
-  };
-}
