--- apps/docs/app/trees/tree-examples/index.ts
@@ -1,9 +1,0 @@
-export { A11ySection } from './A11ySection';
-export { CustomIconsSection } from './CustomIconsSection';
-export { DragDropSection } from './DragDropSection';
-export { FlatteningSection } from './FlatteningSection';
-export { GitStatusSection } from './GitStatusSection';
-export { SearchSection } from './SearchSection';
-export { ThemingSection } from './ThemingSection';
-export { StylingSection } from './StylingSection';
-export { VirtualizationSection } from './VirtualizationSection';
